add guest/host self assignment when entering localnifi mode from SnemulDS.
asap the localnifi mode is chosen, it will actively look for another DS. If the other DS is found, depending on current DS and external DS time, the oldest timestamp is guest, the newest is host.

also when leaving localnifi mode, the nifi is reset on the other connected DS as well.