Rewritten snemul.cfg reload logic. It's removed from repository and it's automatically generated from scratch & it's still persistent in filesystem.
Note: SnemulDS will detect if a valid snemul.cfg exists, and if it's invalid, it'll regenerate one from scratch every time it boots up.
Optionally, you can rebuild it from the "Options" tab, through the button "Config: Reset now" anytime.
The reason of this is that sometimes an invalid snemul.cfg may cause known games to white screens, if so, you can rebuild it immediately without touching a PC.